+++ DeepPhysX_Torch/UNet/UNetConfig.py
@@ -48,7 +48,7 @@ class UNetConfig(TorchNetworkConfig):
                  network_name: str = "UNetNetwork",
                  which_network: int = 0,
                  save_each_epoch: bool = False,
-                 data_type: str = 'float64',
+                 data_type: str = 'float32',
                  lr: Optional[float] = None,
                  require_training_stuff: bool = True,
                  loss: Optional[Any] = None,
